Detector - Number of users
Added number of users consideration on impact calculation

+++ detector/mabsed/mabsed.py
@@ -9,6 +9,7 @@ import numpy as np
 import stats as st
 
 #saving
+import sys # Borrar esto despues de las pruebas
 import os
 import json
 import csv
@@ -36,25 +37,56 @@ class MABSED:
         basic_events = self.phase1()
         return self.phase2(basic_events)
 
+    '''
+        Devuelve una lista que contendra una tupla para cada palabra del vocabulario.
+        Esta tupla describe el impacto que ha tenido dicha palabra en cuanto a numero de menciones
+        ha provocado y numero de usuarios han hablado de ella.
+    '''
     def phase1(self):
         print('Phase 1...')
-        basic_events = []
+        basic_events = [] # Esto va a ser una lista de tuplas
         for vocabulary_entry in self.corpus.vocabulary.items():
-            basic_events.append(self.maximum_contiguous_subsequence_sum(vocabulary_entry))
+            mention_event = self.get_mention_subsequence(vocabulary_entry)
+            # frequency_event = self.get_frequency_subsequence(vocabulary_entry)
+            user_event = self.get_user_subsequence(vocabulary_entry)
+            basic_events.append(self.merge_basic_event(mention_event, user_event))           
         print('   Detected events: %d' % len(basic_events))
         return basic_events
 
-    def maximum_contiguous_subsequence_sum(self, vocabulary_entry):
-        mention_freq = self.corpus.mention_freq[vocabulary_entry[1], :].toarray()
-        mention_freq = mention_freq[0, :]
-        total_mention_freq = np.sum(mention_freq)
+    def merge_basic_event(self, mention_event, user_event):
+        total_magnitude = mention_event[0] + user_event[0]
+        magnitudes = (total_magnitude, mention_event[0], user_event[0])
+        a = min(mention_event[1][0], user_event[1][0])
+        b = max(mention_event[1][1], user_event[1][1])
+        interval = (a, b)
+        total_anomaly = []
+        for i in range(0, self.corpus.time_slice_count):
+            sumatory = mention_event[3][i] + user_event[3][i]
+            total_anomaly.append(sumatory)
+        anomalies = (total_anomaly, mention_event[3], user_event[3])
+        return (magnitudes, interval, mention_event[2], anomalies)
+
+    '''
+        Metodo que se usara en todas las palabras del vocabulario, y para cada una calcula la
+        magnitud del impacto que esta ha generado calculada respecto al numero de menciones
+        que ha provocado.
+        Devuelve una tupla con: (impacto, intervalo, palabra, lista de anomalias)
+    '''
+    def get_mention_subsequence(self, vocabulary_entry):
+        # vocabulary_entry[1] es el puesto en menciones de cada palabra del diccionario, es decir, 0, 1, 2...
+        mention_freq = self.corpus.mention_freq[vocabulary_entry[1], :].toarray() # Aqui tenemos un array con otro array dentro
+        mention_freq = mention_freq[0, :] # Ahora tenemos un unico array, que es la columna de la matriz que corresponde a esa palabra del vocabulario, y cada fila sera un time_slice
+        total_mention_freq = np.sum(mention_freq) # Sumamos para saber cuantos tweets con mencion tiene esa palabra en total en todo el corpus
 
         # compute the time-series that describes the evolution of mention-anomaly
-        anomaly = []
+        anomaly = [] # Aqui tendremos un array cuyo indice seran los time-slices y cuyos valores seran un float que indica lo que sobresale o decaen las menciones asociadas a dicha palabra respecto de lo esperado en ese time_slice
+        # i es el time_slice de cada vuelta
         for i in range(0, self.corpus.time_slice_count):
-            anomaly.append(self.anomaly(i, mention_freq[i], total_mention_freq))
+            # mention_freq[i] es cuantas menciones tiene la palabra en ese time_slice, y es la anomalia observada de esa palabra
+            anomaly.append(self.mention_anomaly(i, mention_freq[i], total_mention_freq))
         max_ending_here = max_so_far = 0
         a = b = a_ending_here = 0
+        # Cuando acabe este bucle for, a y b representaran el intervalo en el que la suma de las anomalias ha destacado mas de la media
         for idx, ano in enumerate(anomaly):
             max_ending_here = max(0, max_ending_here + ano)
             if max_ending_here == 0:
@@ -67,16 +99,103 @@ class MABSED:
                 b = idx
 
         # return the event description
-        max_interval = (a, b)
-        mag = np.sum(anomaly[a:b+1])
-        basic_event = (mag, max_interval, vocabulary_entry[0], anomaly)
+        max_interval = (a, b) # Intervalo en el que la suma de las anomalias ha destacado mas de la media
+        mag = np.sum(anomaly[a:b+1]) # Magnitud de la anomalia
+        basic_event = (mag, max_interval, vocabulary_entry[0], anomaly) # Esto es una tupla, es como una lista pero no podemos editar sus elementos
         return basic_event
+
+    # Devuelve en formato float un numero que indica como sobresalen o decaen las menciones asociadas a esa palabra en el time_slice dado respecto a la media de esa palabra en el corpus
+    def mention_anomaly(self, time_slice, observation, total_mention_freq):
+        # compute the expected frequency of the given word at this time-slice
+        expectation = float(self.corpus.tweet_count[time_slice]) * (float(total_mention_freq)/(float(self.corpus.size)))
+
+        # return the difference between the observed frequency and the expected frequency
+        return observation - expectation
+
+    def get_frequency_subsequence(self, vocabulary_entry):
+        # vocabulary_entry[1] es el puesto en menciones de cada palabra del diccionario, es decir, 0, 1, 2...
+        frequency = self.corpus.global_freq[vocabulary_entry[1], :].toarray() # Aqui tenemos un array con otro array dentro
+        frequency = frequency[0, :] # Ahora tenemos un unico array, que es la columna de la matriz que corresponde a esa palabra del vocabulario, y cada fila sera un time_slice
+        total_freq = np.sum(frequency) # Sumamos para saber la frecuencia de esa palabra en total en todo el corpus
+
+        # compute the time-series that describes the evolution of mention-anomaly
+        anomaly = [] # Aqui tendremos un array cuyo indice seran los time-slices y cuyos valores seran un float que indica lo que sobresale o decae la frecuencia asociada a dicha palabra respecto de lo esperado en ese time_slice
+        # i es el time_slice de cada vuelta
+        for i in range(0, self.corpus.time_slice_count):
+            anomaly.append(self.frequency_anomaly(i, frequency[i], total_freq))
+        max_ending_here = max_so_far = 0
+        a = b = a_ending_here = 0
+        # Cuando acabe este bucle for, a y b representaran el intervalo en el que la suma de las anomalias ha destacado mas de la media
+        for idx, ano in enumerate(anomaly):
+            max_ending_here = max(0, max_ending_here + ano)
+            if max_ending_here == 0:
+                # a new bigger sum may start from here
+                a_ending_here = idx
+            if max_ending_here > max_so_far:
+                # the new sum from a_ending_here to idx is bigger
+                a = a_ending_here+1
+                max_so_far = max_ending_here
+                b = idx
+
+        # return the event description
+        max_interval = (a, b) # Tupla: intervalo en el que la suma de las anomalias ha destacado mas de la media
+        mag = np.sum(anomaly[a:b+1]) # Magnitud de la anomalia
+        basic_event = (mag, max_interval, vocabulary_entry[0], anomaly) # Esto es una tupla, es como una lista pero no podemos editar sus elementos
+        return basic_event
+
+    # Devuelve en formato float un numero que indica como sobresalen o decaen la frecuencia de esa palabra en el time_slice dado respecto a la media de esa palabra en el corpus
+    def frequency_anomaly(self, time_slice, observation, total_freq):
+        # compute the expected frequency of the given word at this time-slice
+        expectation = float(self.corpus.tweet_count[time_slice]) * (float(total_freq)/(float(self.corpus.size)))
+
+        # return the difference between the observed frequency and the expected frequency
+        return observation - expectation
+
+    def get_user_subsequence(self, vocabulary_entry):
+        # vocabulary_entry[1] es el puesto en menciones de cada palabra del diccionario, es decir, 0, 1, 2...
+        user_freq = self.corpus.user_freq[vocabulary_entry[1], :].toarray() # Aqui tenemos un array con otro array dentro
+        user_freq = user_freq[0, :] # Ahora tenemos un unico array, que es la columna de la matriz que corresponde a esa palabra del vocabulario, y cada fila sera un time_slice
+        total_user_freq = np.sum(user_freq) # Sumamos para saber cuantos usuarios estaban hablando de esa palabra en total en todo el corpus
+
+        # compute the time-series that describes the evolution of mention-anomaly
+        anomaly = [] # Aqui tendremos un array cuyo indice seran los time-slices y cuyos valores seran un float que indica lo que sobresale o decaen las menciones asociadas a dicha palabra respecto de lo esperado en ese time_slice
+        # i es el time_slice de cada vuelta
+        for i in range(0, self.corpus.time_slice_count):
+            # mention_freq[i] es cuantas menciones tiene la palabra en ese time_slice, y es la anomalia observada de esa palabra
+            anomaly.append(self.user_anomaly(i, user_freq[i], total_user_freq))
+        max_ending_here = max_so_far = 0
+        a = b = a_ending_here = 0
+        # Cuando acabe este bucle for, a y b representaran el intervalo en el que la suma de las anomalias ha destacado mas de la media
+        for idx, ano in enumerate(anomaly):
+            max_ending_here = max(0, max_ending_here + ano)
+            if max_ending_here == 0:
+                # a new bigger sum may start from here
+                a_ending_here = idx
+            if max_ending_here > max_so_far:
+                # the new sum from a_ending_here to idx is bigger
+                a = a_ending_here+1
+                max_so_far = max_ending_here
+                b = idx
+
+        # return the event description
+        max_interval = (a, b) # Intervalo en el que la suma de las anomalias ha destacado mas de la media
+        mag = np.sum(anomaly[a:b+1]) # Magnitud de la anomalia
+        basic_event = (mag, max_interval, vocabulary_entry[0], anomaly) # Esto es una tupla, es como una lista pero no podemos editar sus elementos
+        return basic_event
+
+    # Devuelve en formato float un numero que indica como sobresalen o decaen los usuarios que hablan sobre esa palabra en el time_slice dado respecto a la media de esa palabra en el corpus
+    def user_anomaly(self, time_slice, observation, total_user_freq):
+        # compute the expected frequency of the given word at this time-slice
+        expectation = float(self.corpus.tweet_count[time_slice]) * (float(total_user_freq)/(float(self.corpus.size)))
+
+        # return the difference between the observed frequency and the expected frequency
+        return observation - expectation
 
     def phase2(self, basic_events):
         print('Phase 2...')
 
         # sort the events detected during phase 1 according to their magnitude of impact
-        basic_events.sort(key=lambda tup: tup[0], reverse=True)
+        basic_events.sort(key=lambda tup: tup[0][0], reverse=True)
 
         # create the event graph (directed) and the redundancy graph (undirected)
         self.event_graph = nx.DiGraph(name='Event graph')
@@ -89,7 +208,7 @@ class MABSED:
         while unique_events < self.k and i < len(basic_events):
             basic_event = basic_events[i]
             main_word = basic_event[2]
-            candidate_words = self.corpus.cooccurring_words(basic_event, self.p)
+            candidate_words = self.corpus.cooccurring_words(basic_event, self.p) # Devuelve una lista con p (10) palabras como maximo
             main_word_freq = self.corpus.global_freq[self.corpus.vocabulary[main_word], :].toarray()
             main_word_freq = main_word_freq[0, :]
             related_words = []
@@ -115,13 +234,6 @@ class MABSED:
         # merge redundant events and save the result
         self.events = self.merge_redundant_events(refined_events)
 
-    def anomaly(self, time_slice, observation, total_mention_freq):
-        # compute the expected frequency of the given word at this time-slice
-        expectation = float(self.corpus.tweet_count[time_slice]) * (float(total_mention_freq)/(float(self.corpus.size)))
-
-        # return the difference between the observed frequency and the expected frequency
-        return observation - expectation
-
     def update_graphs(self, event):
         redundant = False
         main_word = event[2]
@@ -138,7 +250,7 @@ class MABSED:
                         redundant = True
                         break
         if not redundant:
-            self.event_graph.add_node(event[2], interval=event[1], mag=event[0], main_term=True)
+            self.event_graph.add_node(event[2], interval=event[1], mag=event[0][0], main_term=True)
             for related_word, weight in event[3]:
                 self.event_graph.add_edge(related_word, event[2], weight=weight)
         return not redundant
@@ -174,6 +286,7 @@ class MABSED:
                 related_words = event[3]
             else:
                 related_words = self.merge_related_words(main_term, descriptions)
+            related_words.sort(key=lambda tup: tup[1], reverse=True)
             final_event = (event[0], event[1], main_term, related_words, event[4])
             final_events.append(final_event)
         return final_events
@@ -205,24 +318,34 @@ class MABSED:
         for event in self.events:
             self.print_event(event)
 
-    def save_event(self, event, event_counter, output_file):
+    def save_event(self, event, event_id, output_file):
         saving_data = {}
-        saving_data['id'] = event_counter
-        saving_data['impact'] = event[0]
+        saving_data['id'] = event_id
+        impact_json = {
+            'total': float("{0:.2f}".format(event[0][0])),
+            'mention': float("{0:.2f}".format(event[0][1])),
+            'user': float("{0:.2f}".format(event[0][2]))
+        } # Formateamos los impactos a un float de dos decimales
+        saving_data['impact'] = impact_json
         saving_data['main_words'] = event[2].split(', ')
-        saving_data['related_words'] = [None] * len(event[3])
-        related_words_loop = 0
-        for related_word, weight in event[3]: # En la posicion 3 del array event se guarda la lista de palabras relacionadas
-            word_json = {}
-            word_json['word'] = related_word
-            formatted_weight = float("{0:.2f}".format(weight)) # Formateamos el peso a un decimal de dos digitos
-            word_json['weight'] = formatted_weight
-            saving_data['related_words'][related_words_loop] = word_json
-            related_words_loop += 1
-        duration = {}
-        duration['from'] = str(self.corpus.to_date(event[1][0]))
-        duration['to'] = str(self.corpus.to_date(event[1][1]))
-        saving_data['duration'] = duration
+        saving_data['related_words'] = []
+        for word, weight in event[3]:
+            word_json = {
+                'word': word,
+                'weight': float("{0:.2f}".format(weight))
+            } 
+            saving_data['related_words'].append(word_json)
+        duration_json = {
+            'from': str(self.corpus.to_date(event[1][0])),
+            'to': str(self.corpus.to_date(event[1][1]))
+        }
+        saving_data['duration'] = duration_json
+        anomaly_json = {
+            'total': event[4][0],
+            'mention': event[4][1],
+            'user': event[4][2]
+        }
+        saving_data['anomaly'] = anomaly_json
         
         saving_json = json.dumps(saving_data, ensure_ascii=False)
         with open(output_file, 'a') as f:
@@ -230,10 +353,8 @@ class MABSED:
             f.write('\n')
 
     def save_events(self, output_file):
-        try:
+        if os.path.exists(output_file):
             os.remove(output_file)
-        except OSError:
-            pass
 
         for idx, event in enumerate(self.events):
             self.save_event(event, idx+1, output_file)
@@ -253,7 +374,7 @@ class MABSED:
                         if(coordinate != 'null'):
                             points = coordinate.replace('[','').replace(']','').split(",")
                             saving_tweet['lat'] = float(points[1])
-                            saving_tweet['long'] = float(points[0])
+                            saving_tweet['lon'] = float(points[0])
                         saving_json = json.dumps(saving_tweet, ensure_ascii=False)
                         with open(output_file, 'a') as f:
                             f.write(saving_json)
@@ -261,12 +382,9 @@ class MABSED:
                         break
                 break
 
-
     def save_tweets(self, output_file):
-        try:
+        if os.path.exists(output_file):
             os.remove(output_file)
-        except OSError:
-            pass
 
         saving_tweets = {}
         for idx, event in enumerate(self.events):
